Calculate Lag
This file contains some of the research and reasoning behind how I calculated the lag period as well as an algorithm to implement it. I ran the algorithm on all of the different states to get the two metrics (date 50% of cases/ deaths reached as well as difference in the intercepts) for each county. If a certain county didn't have enough data, then their lag was assigned to be 0. This data is all in another .csv file.  Also ran the algorithm on country-wide data (US, UK, France, Spain & Italy) from another data file to see how well the logistic regression fit their case and death data and then calculate their country's average lag period metric.